Refuse an all-skip gate, fail the probe on variation, keep the Qwen add in bf16
The FSDP2 gate skips every arm when it cannot see two GPUs and an
all-skip pytest exits 0, so the runner now refuses to start with fewer
than two. The combine probe exits nonzero when combine_routed_rows
varies between launches instead of only printing it. On the Qwen swap
path the routed rows are still summed in float32 in a fixed order and
rounded once, and the shared expert is added in the activation dtype as
the DeepEP branch does; the former scatter_add accumulated everything,
shared expert included, in bf16 atomics, and that change is now stated
at the call site. The order test's docstring says which four of the 24
orders coincide.

+++ skyrl-train/skyrl_train/models/layers/moe.py
@@ -548,7 +548,10 @@ class MoE(nn.Module):
             selected_experts_indices,
             self.reorderer,
         )
-        out = combine_routed_rows(routed_output, token_indices, x.shape[0], self.top_k)
+        # The routed rows are summed in float32 in a fixed order and rounded once; the shared expert
+        # is then added in the activation dtype, as the DeepEP branch does. The former scatter_add
+        # accumulated everything, shared expert included, in bf16 atomics.
+        out = combine_routed_rows(routed_output, token_indices, x.shape[0], self.top_k).to(x.dtype)
 
         if self.shared_expert is not None:
             shared_output = self.shared_expert(x)
@@ -556,7 +559,7 @@ class MoE(nn.Module):
                 shared_output = F.sigmoid(self.shared_expert_gate(x)) * shared_output
             out = out + shared_output
 
-        return out.to(x.dtype).reshape(bs, slen, dim)
+        return out.reshape(bs, slen, dim)
 
     def init_weights(self, init_std: float):
         self.experts.init_weights(init_std)
